fixing review
and waveclus pre post standard
